Manual

FourierSeriesEvaluators.AbstractFourierSeries — Type

AbstractFourierSeries{N,T,iip}

A supertype for multidimensional Fourier series objects. Given f::AbstractFourierSeries, you can evaluate at a point x with f(x), where x is a vector (or scalar if f is 1d).

Fourier series are periodic maps $T^N \to V$ where T is a real number and $V$ is any vector space. Typically, a Fourier series can be represented by N-dimensional arrays whose elements belong to the vector space. If iip is true, then $V$ is assumed to have mutable elements and inplace array operations are used. Otherwise, $V$ is assumed to be immutable. The period of the series should be specified by values of type T, although no restriction is placed on the inputs to the series, e.g. arguments of type Complex{T} are OK. Additionally, if the caller wants to determine the floating-point precision of the Fourier coefficients, T and the arguments must both have that precision.

FourierSeriesEvaluators.DerivativeSeries — Method

DerivativeSeries{O}(f::FourierSeries)

Construct an evaluator of a Fourier series and all of its derivatives up to order O, which must be a positive integer. O=1 gives the gradient, O=2 gives the Hessian, and so on. The derivatives are returned in order as a tuple (f(x), df(x), d2f(x), ..., dOf(x)) where the entry of order O is given by:

O=0: f

O=1: (dfdx1, ..., dfdxN)

O=2: ((d2fdx1dx1, ..., d2fdx1dxN), ..., (d2fdxNdxN,))

O=3: (((d3fdx1dx1dx1, ..., d3fdx1dx1dxN), ..., (d3fdx1dxNdxN,)), ..., ((d3fdxNdxNdxN,),))

and so on. The fewest number of contractions are made to compute all derivatives. As can be seen from the pattern above, the O-th derivative with partial derivatives i = [a_1 ≤ ... ≤ a_N] is stored in ds(x)[O+1][i[1]][i[2]]...[i[N]]. These indices are given by the simplical generalization of triangular numbers. For examples of how to index into the solution see the unit tests.

FourierSeriesEvaluators.FourierSeries — Type

FourierSeries(coeffs::AbstractArray, [N]; period=2pi, offset=0, deriv=0, shift=0)

Construct a Fourier series whose coefficients are given by the coefficient array array coeffs whose elements should support addition and scalar multiplication, This type represents the Fourier series

\[f(\vec{x}) = \sum_{\vec{n} \in \mathcal I} C_{\vec{n}} \exp(i2\pi\vec{k}_{\vec{n}}\cdot\overrightarrow{x})\]

The indices $\vec{n}$ are the CartesianIndices of coeffs. Also, the keywords, which can either be a single value applied to all dimensions or a collection describing each dimension mean

period: The periodicity of the Fourier series. Equivalent to $2\pi/k$

offset: An offset in the phase index, which must be integer

deriv: The degree of differentiation, implemented as a Fourier multiplier

shift: A translation q such that the evaluation point x is shifted to x-q

If the optional argument N is set, it fixes the number of variables of the Fourier series, which may be less than or equal to ndims(coeffs), and the series is evaluated inplace.

FourierSeriesEvaluators.FourierWorkspace — Type

FourierWorkspace

A workspace for storing the intermediate arrays used during Fourier series evaluations. All functionality is provided by the workspace_allocate, workspace_contract!, and workspace_evaluate routines.

FourierSeriesEvaluators.ManyFourierSeries — Type

ManyFourierSeries(fs::AbstractFourierSeries{N,T,iip}...; period) where {N,T,iip}

Represents a tuple of Fourier series of the same dimension and contracts them all simultaneously. All the series are required to be either inplace or not.

FourierSeriesEvaluators.allocate — Function

allocate(f::AbstractFourierSeries{N}, x, ::Val{d}) where {N,d}

Return a cache that can be used by fourier_contract! to store the result of contracting the coefficients of f along axis d using an input x.

FourierSeriesEvaluators.cis_inv — Method

Return cis and its inverse

FourierSeriesEvaluators.contract! — Function

contract!(cache, f::AbstractFourierSeries{N}, x, ::Val{d}) where {N,d}

Return another Fourier series of dimension N-1 by summing over dimension d of f with the phase factors evaluated at x and using the storage in cache created by a call to fourier_allocate

FourierSeriesEvaluators.evaluate! — Function

evaluate!(cache, f::AbstractFourierSeries{1}, x)

Evaluate the Fourier series at the point x, optionally with cache for inplace evaluation created by a call to fourier_allocate

FourierSeriesEvaluators.fourier_allocate — Method

fourier_allocate(C, x, k, a, ::Val{dim})

Allocate an array of the correct type for contracting the Fourier series along axis dim.

FourierSeriesEvaluators.fourier_contract! — Method

fourier_contract!(r::AbstractArray{T,N-1}, C::AbstractArray{T,N}, x, [k=1, a=0, shift=0, dim=Val(N)]) where {T,N}

Contract dimension dim of array C and write it to the array r, whose axes must match C's (excluding dimension dim). This function uses the indices in axes(C,N) to evaluate the phase factors, which makes it compatible with OffsetArrays as inputs. Optionally, a shift can be provided to manually offset the indices. Also, a represents the order of derivative of the series and must be a Number. The formula for what this routine calculates is:

\[r_{i_{1},\dots,i_{N-1}} = \sum_{i_{N}\in\text{axes}(C,N)} C_{i_{1},\dots,i_{N-1},i_{N}+m+1} (ik (i_{N} + \text{shift}))^{a} \exp(ik x (i_{N} + \text{shift}))\]

FourierSeriesEvaluators.fourier_contract — Method

fourier_contract(C::Vector, x, [k=1, a=0, shift=0, dim=Val(N)])

Identical to fourier_contract! except that it allocates its output.

FourierSeriesEvaluators.fourier_evaluate — Method

fourier_evaluate(C::AbstractArray{T,N}, x::NTuple{N}, [k=1, a=0, shift=0]) where {T,N}

Evaluates a N-D Fourier series C. This function uses the indices in axes(C) to evaluate the phase factors, which makes it compatible with OffsetArrays as inputs. Optionally, a shift can be provided to manually offset the indices. Also, a represents the order of derivative of the series and must be a Number. The arguments x, k, a, shift must all be tuples of length N, the same as the array dimension. The 1-D formula for what this routine calculates is:

\[r = \sum_{i_{\in\text{axes}(C,1)} C_{i} (ik (i + \text{shift}))^{a} \exp(ik x (i + \text{shift}))\]

Multi-dimensional performance hit

This routine is allocation-free, but using it for multidimensional evaluation can be slower than allocating because it always computes the Fourier coefficients on the fly. Thus, it is typically more efficient to compute the outermost dimensions of the series with fourier_contract! and then use this routine for the innermost dimension, which is faster because it doesn't use inplace operations. FourierSeries implements this behavior.

FourierSeriesEvaluators.frequency — Function

frequency(f::AbstractFourierSeries, [dim]) = map(inv, period(f, [dim]))

Return a tuple containing the frequency, or inverse of the period, of f. Optionally you can specify a dimension to just get the frequency of that dimension.

FourierSeriesEvaluators.period — Function

period(f::AbstractFourierSeries, [dim])

Return a tuple containing the periodicity of f. Optionally you can specify a dimension to just get the period of that dimension. This should have the floating-point precision of the input used for the Fourier series evaluation.

FourierSeriesEvaluators.workspace_allocate — Method

workspace_allocate(s::AbstractFourierSeries{N}, x::NTuple{N}, [len::NTuple{N}=ntuple(one,N)])

Allocates a FourierWorkspace for the Fourier series s that can be used to evaluate the series multiple times without allocating on-the-fly. The len argument can indicate how many copies of workspace should be made for each variable for downstream use in parallel workloads.

The workspace is constructed recursively starting from the outer dimension and moving towards the inner dimension so as to access memory contiguously. Thus, the outer dimension has len[N] workspace copies and each of these has len[N-1] workspaces for the next variable. In total there are prod(len) leaf-level caches to use for parallel workloads.

FourierSeriesEvaluators.workspace_contract! — Function

workspace_contract!(ws, x, [i=1])

Returns a workspace with the series contracted at variable x in the outer dimension. The index i selects which workspace in the cache to assign the new data.

FourierSeriesEvaluators.workspace_evaluate! — Function

workspace_evaluate!(ws, x, [i=1])

Return the 1-d series evaluated at the variable x, using cache sector i.

FourierSeriesEvaluators.workspace_evaluate — Method

workspace_evaluate(ws, x)

Evaluates the series using the workspace.