Manual

FourierSeriesEvaluators.AbstractFourierSeries — Type

AbstractFourierSeries{N,T}

A supertype for Fourier series that are periodic maps $\R^N \to V$ where $V$ is any vector space with elements of type T. Typically these can be represented by N-dimensional arrays whose elements belong to the vector space.

(f::AbstractFourierSeries)(x) = evaluate(f, promote(x...))

Evaluate the Fourier series at the given point (see evaluate).

FourierSeriesEvaluators.AbstractInplaceFourierSeries — Type

AbstractInplaceFourierSeries{N,T} <: AbstractFourierSeries{N,T}

A supertype for Fourier series evaluated in place. These define the contract! method instead of contract.

FourierSeriesEvaluators.FourierSeries — Type

FourierSeries(coeffs::AbstractArray; period=2pi, offset=0, deriv=0, shift=0)

Construct a Fourier series whose coefficients are given by the coefficient array array coeffs whose elements should support addition and scalar multiplication, This type represents the Fourier series

\[f(\vec{x}) = \sum_{\vec{n} \in \mathcal I} C_{\vec{n}} \exp(i2\pi\vec{k}_{\vec{n}}\cdot\overrightarrow{x})\]

The indices $\vec{n}$ are the CartesianIndices of coeffs. Also, the keywords, which can either be a single value applied to all dimensions or a collection describing each dimension mean

period: The periodicity of the Fourier series. Equivalent to $2\pi/k$

offset: An offset in the phase index, which must be integer

deriv: The degree of differentiation, implemented as a Fourier multiplier

shift: A translation q such that the evaluation point x is shifted to x-q

FourierSeriesEvaluators.InplaceFourierSeries — Type

InplaceFourierSeries(coeffs::AbstractArray; period=2pi, offset=0, deriv=0, shift=0)

Similar to FourierSeries except that it doesn't allocate new arrays for every call to contract and contract is limited to the outermost dimension/variable of the series.

FourierSeriesEvaluators.ManyFourierSeries — Type

ManyFourierSeries(fs::AbstractFourierSeries{N}...) where {N}

Represents a tuple of Fourier series of the same dimension and periodicity and contracts them all simultaneously.

Base.eltype — Method

eltype(::AbstractFourierSeries{N,T}) where {N,T}

Returns T, the type of the input data to the Fourier series. For the output type, see fourier_type

FourierSeriesEvaluators.cis_inv_op — Method

cis_inv_op(x::Type, k::Type)

Return an expression with a function to invert cis(x*k) based on the type of x*k. When k*x isa Real then cis(x*k) is a root of unity whose inverse is its complex conjugate.

FourierSeriesEvaluators.coefficients — Method

coefficients(f::AbstractFourierSeries) = nothing

Return the underlying array representing the Fourier series, if applicable. There are cases with multiple Fourier series where this many not make sense.

FourierSeriesEvaluators.contract — Function

contract(f::AbstractFourierSeries{N}, x::Number, ::Val{dim}) where {N,dim}

Return another Fourier series of dimension N-1 by summing over dimension dim of f with the phase factors evaluated at x.

For developers

Implementations of the interface need to provide an implementation of contract with the same signature as the above that specialize on the concrete type of f. While any value of dim from 1 to N is can be implemented, it is most important to implement dim==N.

FourierSeriesEvaluators.contract! — Function

contract!(f::AbstractInplaceFourierSeries, x::Number, dim::Type{Val{d}})

An in-place version of contract, however the argument dim must be a Val{d} in order to dispatch to the specific contract method. This should return f.

For developers

An AbstractInplaceFourierSeries only needs to implement contract!, which is set up to be called by contract

FourierSeriesEvaluators.evaluate — Function

evaluate(f::AbstractFourierSeries, x)

Evaluate the Fourier series at the point x. By default x is wrapped into a tuple and the Fourier series is contracted along the outer dimension.

For developers

Implementations of the interface only need to define a method specializing on the concrete type T of f::T with signature evaluate(T, ::NTuple{1}) while evaluation of the other dimensions can be delegated to contract.

FourierSeriesEvaluators.fourier_contract! — Method

fourier_contract!(r::AbstractArray{T,N-1}, C::AbstractArray{T,N}, x, [k=1, a=0, shift=0, dim=Val(N)]) where {T,N}

Contract dimension dim of array C and write it to the array r, whose axes must match C's (excluding dimension dim). This function uses the indices in axes(C,N) to evaluate the phase factors, which makes it compatible with OffsetArrays as inputs. Optionally, a shift can be provided to manually offset the indices. Also, a represents the order of derivative of the series (see fourier_multiplier for optimized values). The formula for what this routine calculates is:

\[r_{i_{1},\dots,i_{N-1}} = \sum_{i_{N}\in\text{axes}(C,N)} C_{i_{1},\dots,i_{N-1},i_{N}+m+1} (ik (i_{N} + \text{shift}))^{a} \exp(ik x (i_{N} + \text{shift}))\]

FourierSeriesEvaluators.fourier_contract — Method

fourier_contract(C::Vector, x, [k=1, a=0, shift=0, dim=Val(N)])

Identical to fourier_contract! except that it allocates its output.

FourierSeriesEvaluators.fourier_evaluate — Function

fourier_evaluate(C::AbstractVector, x, [k=1, a=0, shift=0])

Evaluates a 1D Fourier series C. This function uses the indices in axes(C,1) to evaluate the phase factors, which makes it compatible with OffsetArrays as inputs. Optionally, a shift can be provided to manually offset the indices. Also, a represents the order of derivative of the series (see fourier_multiplier for optimized values). The formula for what this routine calculates is:

\[r = \sum_{i_{\in\text{axes}(C,1)} C_{i} (ik (i + \text{shift}))^{a} \exp(ik x (i + \text{shift}))\]

FourierSeriesEvaluators.fourier_multiplier — Method

fourier_multiplier(a::Type, indices::Expr...)

Transform the index expressions indices to be exponentiated to obtain an efficient Fourier multiplier that can be interpolated into the source code.

Optimizations of the Fourier multiplier(in the following let i be an index):

a === Val{0}: returns :nothing expressions since there is no derivative

a === Val{1}: returns i to avoid exponentiation

a <: Integer: returns i^a to use integer exponent

Otherwise the fallback Fourier multiplier is Complex(i)^a

FourierSeriesEvaluators.fourier_type — Method

fourier_type(::Type{T}, x) where T = Base.promote_op(*, T, phase_type(x))
fourier_type(C::AbstractFourierSeries, x) = fourier_type(eltype(f), x)

Returns the output type of the Fourier series.

FourierSeriesEvaluators.nref_dim — Method

nref_dim(N::Int, C::Symbol, i::Symbol, dim::Int, j::Union{Symbol,Expr})

Return an N dimensional indexing expression like C[i_1,i_2,j] or C[j,i_1,i_2] where N controls the total number of dimensions, dim controls the position of the inserted index j.

FourierSeriesEvaluators.period — Function

period(f::AbstractFourierSeries{N}) where {N}

Return a NTuple{N} whose m-th element corresponds to the period of f along its m-th input dimension. Typically, these values set the units of length for the problem.

FourierSeriesEvaluators.phase_type — Method

phase_type(x) = Base.promote_op(cis, eltype(x))

Returns the type of exp(im*x).